Added new components
